build fix
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@30397 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/PGMSavedState.cpp
@@ -1894,8 +1894,8 @@ static DECLCALLBACK(int)  pgmR3LiveVote(PVM pVM, PSSMHANDLE pSSM, uint32_t uPass
      * Come up with a completion percentage.  Currently this is a simple
      * dirty page (long term) vs. total pages ratio + some pass trickery.
      */
-    unsigned uPctDirty = cDirtyPagesLong
-                       / (long double)(pVM->pgm.s.cAllPages - pVM->pgm.s.LiveSave.cIgnoredPages - pVM->pgm.s.cZeroPages);
+    unsigned uPctDirty = (unsigned)(  (long double)cDirtyPagesLong
+                                    / (pVM->pgm.s.cAllPages - pVM->pgm.s.LiveSave.cIgnoredPages - pVM->pgm.s.cZeroPages);
     if (uPctDirty <= 100)
         SSMR3HandleReportLivePercent(pSSM, RT_MIN(100 - uPctDirty, uPass * 2));
     else
